requirements.txt
selenium
gspread>=3.6.0
requests>=2.20.0
GitPython==3.1.11
e-objects>=1.20.2
google-api-python-client==1.10.0
